Create Word_char_frequency.py
Python program to input a paragraph as a String. Display all the words containing the character ‘C’ and order them in ascending order by the frequency (the number of times a particular character occurs in a word) of the character ‘C’. Here ‘C’ is any character that has been taken as user input. Take all the inputs from the user.